Started to make code more modular.
Instead of a "CommandPane" that really was a panel with many tabbed
subpanels, I started to create "Frame"s that are standalone windows
that pop up when you need to do a specific task like connect to the
server, choose a simulation, or change the simulation colors.  The
ReColorFrame is much improved.  Now you can select specific attributes
that you want displayed, and are automatically provided with upper and
lower bounds for those attributes.

Created the SimulationFrame, which now only contains the ParentPanel because
the ParentPanel works, but soon will contain the 4 distinct elements:
1. color bar
2. main view
3. aux. view
4. toolbar