Group selection is starting take shape.  At least the dialog is starting to look a lot more like the Mac mail rule creation dialog so that you can add and subtract criteria.  The next step is to start using the CreateGroup message.  We also need to start drawing bounding boxes for selecting positions.
Graeme changed the choose simulation message so that now we know every
attribute for every family of particles.  This necessitated the creation of
a Family object to hang onto all of the new information.

Recolor finally works reasonably intelligently.  If you only want to deal
with particular particle types, you can.