Added "MapReduce" style particle reduction.  The "reduceParticle" method takes 1) code to map a particle to a tuple, the first item of which is a key. 2) code to combine a list of tuples with the same key.
tipsy.py shows how this function can be used to implement the "boxstat"
command.

There was also significant changes in the way the Worker interacts with the
Python interpreter.  A PythonObjectLocal class is implemented that allows
Python execution without going through the CCS layer.  Worker is no longer
a python array in the ci file.