Changed rotationcurve() to use particle softening value instead of user-input eps_grav and epsgas_grav values. This is contrary to the behavior in TIPSY. Removed related values from config.py
Added acsmooth function to replace initialized array of values.

Made numerous minor edits to trot_cur.py and grav.py.

Function rotationcurve() still takes approximately a minute to
finish when running on one CPU. By selectively disabling code,
the grav function appears to be the culprit. I was not able to
determine a specific behavior of grav which causes this beyond
the existance of several charm.reduceParticle() calls.