Final, definitive fix for Tom Quinn's "green shadows" color table lookup bug.
This is a weird roundoff-sensitive problem with his GeForce 7600GS.  Unlike
other cards of that generation, this model seems to tend to round up textel
coordinates near boundaries.  The fix is to move far from the color table index
boundaries, by scaling by 0.997, 255/256.  This seems to work with other cards too.